zustand작성,오류많음
